fix(notify): ring the bell when an action-aware notify-send is refused
Waiting for the click in a goroutine meant the call reported success the
instant it started, so a daemon that refused the banner lost the ping
entirely: the bell had already been skipped. The call now gives the
banner a second to prove it reached the desktop, and a refusal in that
window comes back as the failure it is. A banner on screen still holds
the reply for the user's click.

The build-tagged stubs lose comments that only restated their bodies, and
the WSL and focus tests move beside the files they cover.
